+++ resources-library/src/commonMain/kotlin/ByteArrayExt.kt
@@ -68,4 +68,3 @@ internal fun ByteArray.decodeUtf16(): String {
         }
     }
 }
-