+++ resources-library/src/commonMain/kotlin/ResourceReadException.kt
@@ -1,6 +1,6 @@
 package com.goncalossilva.resources
 
-public class FileReadException : RuntimeException {
+public class ResourceReadException : RuntimeException {
     public constructor() : super()
     public constructor(message: String?) : super(message)
     public constructor(message: String?, cause: Throwable?) : super(message, cause)